round2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-personal-assistant-conflict-v2"
version = "0.2.0"
description = "Round 2 personalized long-horizon OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.35.0",
    "pydantic>=2.12.0",
    "openai>=2.7.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
]

[project.scripts]
server = "assistant_conflict_env.server_main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
assistant_conflict_env = ["fixtures/*.json"]
